requirements.txt
amqp==5.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
idna==3.7
iniconfig==2.0.0
kombu==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
